requirements.txt
# Web Data Extraction Project - Dependencies
# Install with: pip install -r requirements.txt

# Web Scraping Libraries
selenium==4.16.0
beautifulsoup4==4.12.2
lxml==5.1.0

# Database
mysql-connector-python==8.2.0

# Utilities
requests==2.31.0
urllib3==2.1.0

# Chrome WebDriver Manager (optional - auto-downloads chromedriver)
webdriver-manager==4.0.1

# Web Dashboard & Analytics
streamlit==1.30.0
pandas==2.1.4
plotly==5.18.0
numpy==1.26.3
scipy==1.11.4
